Continue work on inventory translation

+++ src/main/java/net/raphimc/viabedrock/protocol/data/ProtocolConstants.java
@@ -37,7 +37,7 @@ public class ProtocolConstants {
     public static final int BEDROCK_RAKNET_PROTOCOL_VERSION = 11;
     public static final int BEDROCK_DEFAULT_PORT = 19132;
     public static final int BEDROCK_COMMAND_VERSION = 39;
-    public static final short BEDROCK_REQUEST_CHUNK_RADIUS_MAX_RADIUS = 28;
+    public static final byte BEDROCK_REQUEST_CHUNK_RADIUS_MAX_RADIUS = 28;
 
     public static StructuredDataContainer createStructuredDataContainer() {
         final StructuredDataContainer data = new StructuredDataContainer();
